requirements.txt
python-telegram-bot[ext]
fastapi
uvicorn
gspread
oauth2client
python-dotenv
pydantic
httpx<0.27.0